Thêm chức năng Tạo, Cập nhật, Xóa, Đăng ký, Thu hồi đăng kí đề tài.

+++ InternshipManagement/Models/ViewModels/DeTaiRegistrationStatusVm.cs
@@ -2,9 +2,12 @@
 {
     public int MaSv { get; set; }
     public string MaDt { get; set; } = "";
-    public byte? ThisTrangThai { get; set; }          // NULL/0/1/2/3/4
+
+    // đổi sang int? để nhận được -1 từ proc
+    public int? ThisTrangThai { get; set; }           // -1 / 0..5 / null
     public bool HasOtherTopic123 { get; set; }
+
     public string? OtherMaDt { get; set; }
     public string? OtherTenDt { get; set; }
-    public byte? OtherTrangThai { get; set; }
+    public int? OtherTrangThai { get; set; }          // 1/2/3 hoặc null
 }
